pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillkit"
version = "0.1.0"
description = "Python library implementing Anthropic's Agent Skills functionality for LLM-powered agents"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Massimo Olivieri", email = "olivmassimo@gmail.com"}
]
keywords = [
    "anthropic",
    "claude",
    "skills",
    "agents",
    "llm",
    "langchain",
    "ai",
    "prompt-engineering",
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=0.1.0",
    "pydantic>=2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML",
]

[project.urls]
Homepage = "https://github.com/maxvaega/skillkit"
Repository = "https://github.com/maxvaega/skillkit"
Documentation = "https://github.com/maxvaega/skillkit#readme"
Issues = "https://github.com/maxvaega/skillkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/skillkit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: Integration tests with external frameworks (LangChain)",
    "performance: Performance and timing tests",
    "slow: Tests that take >5 seconds to complete",
    "requires_langchain: Tests requiring langchain-core to be installed",
]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=src/skillkit",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=70",
]

[tool.coverage.run]
branch = true
source = ["src/skillkit"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "E402",  # module level import not at top of file
    "UP006", # Use X instead of typing.X for annotations (Python 3.9 compatibility)
    "UP035", # typing.X is deprecated (Python 3.9 compatibility)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",  # Unused function arguments in tests (fixtures)
    "S101", # Use of assert in tests
]

[tool.ruff.lint.isort]
known-first-party = ["skillkit"]
